Handle boolean flags (#30)
Not all options require an argument, but we treated everything (minus leading shorthands) that as needing one. The easiest way to demonstrate is with an example:

```bash
$ git push -f origin master
```

Here, `-f` doesn't take an argument, so returning `{ f: "origin" }` is definitely wrong.

I'm solving this by passing a `boolean` array into the parser. This is so we can support both the current and the boolean syntax:

```bash
$ curl -X POST ...
$ git push -f origin master
```